Move to misc_utils from merlin-models
