+++ transformers4rec/torch/utils/schema_utils.py
@@ -127,3 +127,42 @@ def _get_sparse_tensor(values, indices, num_rows, seq_limit):
     sparse_tensor = torch.sparse_coo_tensor(indices.T, values, torch.Size([num_rows, seq_limit]))
 
     return sparse_tensor.to_dense()
+
+
+def _augment_schema(
+    schema,
+    cats=None,
+    conts=None,
+    labels=None,
+    sparse_names=None,
+    sparse_max=None,
+    sparse_as_dense=False,
+):
+    from merlin.schema import ColumnSchema, Tags
+
+    schema = schema.select_by_name(conts + cats + labels)
+
+    labels = [labels] if isinstance(labels, str) else labels
+    for label in labels or []:
+        schema[label] = schema[label].with_tags(Tags.TARGET)
+    for label in cats or []:
+        schema[label] = schema[label].with_tags(Tags.CATEGORICAL)
+    for label in conts or []:
+        schema[label] = schema[label].with_tags(Tags.CONTINUOUS)
+
+    # Set the appropriate properties for the sparse_names/sparse_max/sparse_as_dense
+    for col in sparse_names or []:
+        cs = schema[col]
+        properties = cs.properties
+        if sparse_max and col in sparse_max:
+            properties["value_count"] = {"max": sparse_max[col]}
+        schema[col] = ColumnSchema(
+            name=cs.name,
+            tags=cs.tags,
+            dtype=cs.dtype,
+            is_list=True,
+            is_ragged=not sparse_as_dense,
+            properties=properties,
+        )
+
+    return schema
